refactor(auth): extract useOtpFlow hook + OtpCodeStep to dedupe login/signup OTP flow (#3184)
